+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/feature/UserSessionScope.kt
@@ -77,8 +77,8 @@ import com.wire.kalium.logic.data.conversation.ProposalTimer
 import com.wire.kalium.logic.data.conversation.SubconversationRepositoryImpl
 import com.wire.kalium.logic.data.conversation.UpdateKeyingMaterialThresholdProvider
 import com.wire.kalium.logic.data.conversation.UpdateKeyingMaterialThresholdProviderImpl
-import com.wire.kalium.logic.data.e2ei.CrlRepository
-import com.wire.kalium.logic.data.e2ei.CrlRepositoryDataSource
+import com.wire.kalium.logic.data.e2ei.CertificateRevocationListRepository
+import com.wire.kalium.logic.data.e2ei.CertificateRevocationListRepositoryDataSource
 import com.wire.kalium.logic.data.e2ei.E2EIRepository
 import com.wire.kalium.logic.data.e2ei.E2EIRepositoryImpl
 import com.wire.kalium.logic.data.event.EventDataSource
@@ -207,8 +207,8 @@ import com.wire.kalium.logic.feature.conversation.mls.OneOnOneResolverImpl
 import com.wire.kalium.logic.feature.debug.DebugScope
 import com.wire.kalium.logic.feature.e2ei.ACMECertificatesSyncWorker
 import com.wire.kalium.logic.feature.e2ei.ACMECertificatesSyncWorkerImpl
-import com.wire.kalium.logic.feature.e2ei.CheckCrlWorker
-import com.wire.kalium.logic.feature.e2ei.CheckCrlWorkerImpl
+import com.wire.kalium.logic.feature.e2ei.CertificateRevocationListCheckWorker
+import com.wire.kalium.logic.feature.e2ei.CertificateRevocationListCheckWorkerImpl
 import com.wire.kalium.logic.feature.e2ei.usecase.CheckRevocationListUseCase
 import com.wire.kalium.logic.feature.e2ei.usecase.CheckRevocationListUseCaseImpl
 import com.wire.kalium.logic.feature.e2ei.usecase.EnrollE2EIUseCase
@@ -1509,8 +1509,8 @@ class UserSessionScope internal constructor(
             userStorage.database.clientDAO,
             userStorage.database.metadataDAO,
         )
-    private val crlRepository: CrlRepository
-        get() = CrlRepositoryDataSource(
+    private val certificateRevocationListRepository: CertificateRevocationListRepository
+        get() = CertificateRevocationListRepositoryDataSource(
             acmeApi = globalScope.unboundNetworkContainer.acmeApi,
             metadataDAO = userStorage.database.metadataDAO,
             userConfigRepository = userConfigRepository
@@ -1527,9 +1527,9 @@ class UserSessionScope internal constructor(
         )
     }
 
-    private val checkCrlWorker: CheckCrlWorker by lazy {
-        CheckCrlWorkerImpl(
-            crlRepository = crlRepository,
+    private val certificateRevocationListCheckWorker: CertificateRevocationListCheckWorker by lazy {
+        CertificateRevocationListCheckWorkerImpl(
+            certificateRevocationListRepository = certificateRevocationListRepository,
             incrementalSyncRepository = incrementalSyncRepository,
             checkRevocationList = checkRevocationList,
         )
@@ -1888,7 +1888,7 @@ class UserSessionScope internal constructor(
 
     private val checkRevocationList: CheckRevocationListUseCase
         get() = CheckRevocationListUseCaseImpl(
-            crlRepository = crlRepository,
+            certificateRevocationListRepository = certificateRevocationListRepository,
             currentClientIdProvider = clientIdProvider,
             mlsClientProvider = mlsClientProvider,
             mLSConversationsVerificationStatusesHandler = mlsConversationsVerificationStatusesHandler
@@ -1927,7 +1927,7 @@ class UserSessionScope internal constructor(
         }
 
         launch {
-            checkCrlWorker.execute()
+            certificateRevocationListCheckWorker.execute()
         }
 
         launch {
